+++ astropy/io/ascii/tests/test_compressed.py
@@ -14,7 +14,7 @@ def test_gzip(filename):
     t_comp = read(os.path.join(ROOT, filename))
     t_uncomp = read(os.path.join(ROOT, filename.replace('.gz', '')))
     assert t_comp.dtype.names == t_uncomp.dtype.names
-    assert np.all(t_comp._data == t_uncomp._data)
+    assert np.all(t_comp.as_array() == t_uncomp.as_array())
 
 
 @pytest.mark.parametrize('filename', ['t/short.rdb.bz2', 't/ipac.dat.bz2'])
@@ -22,4 +22,4 @@ def test_bzip2(filename):
     t_comp = read(os.path.join(ROOT, filename))
     t_uncomp = read(os.path.join(ROOT, filename.replace('.bz2', '')))
     assert t_comp.dtype.names == t_uncomp.dtype.names
-    assert np.all(t_comp._data == t_uncomp._data)
+    assert np.all(t_comp.as_array() == t_uncomp.as_array())