+++ astropy/table/column.py
@@ -54,7 +54,7 @@ class BaseColumn(np.ndarray):
 
     def __new__(cls, data=None, name=None,
                 dtype=None, shape=(), length=0,
-                description=None, unit=None, format=None, meta=None):
+                description=None, unit=None, format=None, meta=None, copy=False):
 
         if data is None:
             dtype = (np.dtype(dtype).str, shape)
@@ -64,7 +64,7 @@ class BaseColumn(np.ndarray):
             # BaseColumn with none of the expected attributes.  In this case
             # do NOT execute this block which initializes from ``data``
             # attributes.
-            self_data = np.asarray(data.data, dtype=dtype)
+            self_data = np.array(data.data, dtype=dtype, copy=copy)
             if description is None:
                 description = data.description
             if unit is None:
@@ -77,12 +77,12 @@ class BaseColumn(np.ndarray):
                 name = data.name
         elif isinstance(data, Quantity):
             if unit is None:
-                self_data = np.asarray(data, dtype=dtype)
+                self_data = np.array(data, dtype=dtype, copy=copy)
                 unit = data.unit
             else:
-                self_data = np.asarray(data.to(unit), dtype=dtype)
+                self_data = np.array(data.to(unit), dtype=dtype, copy=copy)
         else:
-            self_data = np.asarray(data, dtype=dtype)
+            self_data = np.array(data, dtype=dtype, copy=copy)
 
         self = self_data.view(cls)
         self._name = fix_column_name(name)
@@ -205,16 +205,6 @@ class BaseColumn(np.ndarray):
     def __ne__(self, other):
         return self.data.__ne__(other)
 
-    # Set items using a view of the underlying data, as it gives an
-    # order-of-magnitude speed-up. [#2994]
-    def __setitem__(self, index, value):
-        self.data[index] = value
-
-    # Set slices using a view of the underlying data, as it gives an
-    # order-of-magnitude speed-up.  Only gets called in Python 2.  [#3020]
-    def __setslice__(self, start, stop, value):
-        self.data.__setslice__(start, stop, value)
-
     def __array_finalize__(self, obj):
         # Obj will be none for direct call to Column() creator
         if obj is None:
@@ -274,9 +264,6 @@ class BaseColumn(np.ndarray):
         if self.parent_table is not None:
             table = self.parent_table
             table.columns._rename_column(self.name, val)
-            table._data.dtype.names = list(table.columns)
-            if table.masked:
-                table._data.mask.dtype.names = list(table.columns)
 
         self._name = val
 
@@ -623,14 +610,14 @@ class Column(BaseColumn):
 
     def __new__(cls, data=None, name=None,
                 dtype=None, shape=(), length=0,
-                description=None, unit=None, format=None, meta=None):
+                description=None, unit=None, format=None, meta=None, copy=False):
 
         if isinstance(data, MaskedColumn) and np.any(data.mask):
             raise TypeError("Cannot convert a MaskedColumn with masked value to a Column")
 
         self = super(Column, cls).__new__(cls, data=data, name=name, dtype=dtype,
                                           shape=shape, length=length, description=description,
-                                          unit=unit, format=format, meta=meta)
+                                          unit=unit, format=format, meta=meta, copy=copy)
         return self
 
     def __repr__(self):
@@ -654,6 +641,16 @@ class Column(BaseColumn):
         return six.text_type(self).encode('utf-8')
     if six.PY2:
         __str__ = __bytes__
+
+    # Set items using a view of the underlying data, as it gives an
+    # order-of-magnitude speed-up. [#2994]
+    def __setitem__(self, index, value):
+        self.data[index] = value
+
+    # # Set slices using a view of the underlying data, as it gives an
+    # # order-of-magnitude speed-up.  Only gets called in Python 2.  [#3020]
+    def __setslice__(self, start, stop, value):
+        self.data.__setslice__(start, stop, value)
 
     # We do this to make the methods show up in the API docs
     name = BaseColumn.name
@@ -741,7 +738,7 @@ class MaskedColumn(Column, ma.MaskedArray):
 
     def __new__(cls, data=None, name=None, mask=None, fill_value=None,
                 dtype=None, shape=(), length=0,
-                description=None, unit=None, format=None, meta=None):
+                description=None, unit=None, format=None, meta=None, copy=False):
 
         if mask is None and hasattr(data, 'mask'):
             mask = data.mask
@@ -757,7 +754,7 @@ class MaskedColumn(Column, ma.MaskedArray):
         # First just pass through all args and kwargs to BaseColumn, then wrap that object
         # with MaskedArray.
         self_data = BaseColumn(data, dtype=dtype, shape=shape, length=length, name=name,
-                               unit=unit, format=format, description=description, meta=meta)
+                               unit=unit, format=format, description=description, meta=meta, copy=copy)
         self = ma.MaskedArray.__new__(cls, data=self_data, mask=mask)
 
         # Note: do not set fill_value in the MaskedArray constructor because this does not
@@ -793,9 +790,6 @@ class MaskedColumn(Column, ma.MaskedArray):
         """Set fill value both in the masked column view and in the parent table
         if it exists.  Setting one or the other alone doesn't work."""
         val = self._fix_fill_value(val)
-
-        if self.parent_table:
-            self.parent_table._data[self._name].fill_value = val
 
         # Yet another ma bug workaround: If the value of fill_value for a string array is
         # requested but not yet set then it gets created as 'N/A'.  From this point onward
@@ -864,6 +858,15 @@ class MaskedColumn(Column, ma.MaskedArray):
             out.meta = deepcopy(getattr(self, 'meta', {}))
 
         return out
+
+    # Set items and slices using MaskedArray method, instead of falling through
+    # to the (faster) Column version which uses an ndarray view.  This doesn't
+    # copy the mask properly. See test_setting_from_masked_column test.
+    def __setitem__(self, index, value):
+        ma.MaskedArray.__setitem__(self, index, value)
+
+    def __setslice__(self, start, stop, value):
+        ma.MaskedArray.__setslice__(self, start, stop, value)
 
     # We do this to make the methods show up in the API docs
     name = BaseColumn.name